Extract state view projections from orchestrator
